requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
line-bot-sdk==3.22.0
python-dotenv==1.2.2
loguru==0.7.3
langchain-openai==1.1.7